requirements.txt
absl-py==0.7.1
astor==0.8.0
astroid==2.2.5
autopep8==1.4.4
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
isort==4.3.20
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
mock==3.0.5
numpy==1.16.4
Pillow==6.0.0
protobuf==3.8.0
pycodestyle==2.5.0
pylint==2.3.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.1
